SA_Test_main/requirements.txt
fastapi
uvicorn
python-multipart
pandas
Pillow
requests